+++ remote/src/main/java/com/kelsos/mbrc/viewmodels/ConnectionStatusModel.java
@@ -4,13 +4,13 @@ import com.google.inject.Singleton;
 import com.kelsos.mbrc.annotations.Connection;
 
 @Singleton public class ConnectionStatusModel {
-  @Connection.Status private int status;
+  @Connection.Status private long status;
 
-  @Connection.Status public int getStatus() {
+  @Connection.Status public long getStatus() {
     return status;
   }
 
-  public void setStatus(@Connection.Status int status) {
+  public void setStatus(@Connection.Status long status) {
     this.status = status;
   }
 }
